bump package id component to 32 bits
and to make the base64 round even, bump sha256 to 200 bits (up from 192)

+++ src/Package/Manifest.zig
@@ -36,7 +36,7 @@ pub const ErrorMessage = struct {
 };
 
 name: []const u8,
-id: u16,
+id: u32,
 version: std.SemanticVersion,
 version_node: Ast.Node.Index,
 dependencies: std.StringArrayHashMapUnmanaged(Dependency),
@@ -149,7 +149,7 @@ const Parse = struct {
     errors: std.ArrayListUnmanaged(ErrorMessage),
 
     name: []const u8,
-    id: u16,
+    id: u32,
     version: std.SemanticVersion,
     version_node: Ast.Node.Index,
     dependencies: std.StringArrayHashMapUnmanaged(Dependency),
